Display global spinner when loading pre-defined/external benchmark results
